% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\name{colors}
\alias{colors}
\alias{colors_discrete}
\alias{colors_continuous}
\title{Default Colors for Visualization in Package pomdp}
\usage{
colors_discrete(n, col = NULL)

colors_continuous(val, col = NULL)
}
\arguments{
\item{n}{number of states.}

\item{col}{custom color palette. \code{colors_discrete()} uses the first n colors.
\code{colors_continuous()} uses these colors to calculate a palette (see \code{\link[grDevices:colorRamp]{grDevices::colorRamp()}})}

\item{val}{a vector with values to be translated to colors.}
}
\value{
\code{colors_discrete()} returns a color palette and
\code{colors_continuous()} returns the colors associated with the supplied values.
}
\description{
Default discrete and continuous colors used in pomdp for states (nodes), beliefs and values.
}
\examples{
colors_discrete(5)

colors_continuous(runif(10))
}
