\name{solver.output}
\alias{solver.output}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Return output of the POMDP solver
}
\description{
The function simply returns all the output generated by the 'pomdp-solve'
}
\usage{
solver.output(x)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{x}{
object of class POMDP
}
}

\value{
returns a character that contains the output of the 'pomdp-solve' including its iterations
}

\seealso{
\code{\link{solve_POMDP}}
}
\examples{
data("TigerProblem")
sol <- solve_POMDP(model = TigerProblem, grid_size = 10)

solver.output(sol)
}