% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MDP_policy_functions.R
\name{MDP_policy_functions}
\alias{MDP_policy_functions}
\alias{q_values_MDP}
\alias{approx_MDP_policy_evaluation}
\alias{greedy_MDP_action}
\alias{greedy_MDP_policy}
\alias{random_MDP_policy}
\alias{manual_MDP_policy}
\title{Functions for MDP Policies}
\usage{
q_values_MDP(model, U = NULL)

approx_MDP_policy_evaluation(pi, model, U = NULL, k_backups = 10)

greedy_MDP_action(s, Q, epsilon = 0, prob = FALSE)

greedy_MDP_policy(Q)

random_MDP_policy(model, prob = NULL)

manual_MDP_policy(model, actions)
}
\arguments{
\item{model}{an MDP problem specification.}

\item{U}{a vector with value function representing the state utilities
(expected sum of discounted rewards from that point on).
If \code{model} is a solved model, then the state
utilities are taken from the solution.}

\item{pi}{a policy as a data.frame with at least columns for states and action.}

\item{k_backups}{number of look ahead steps used for approximate policy evaluation
used by the policy iteration method.}

\item{s}{a state.}

\item{Q}{an action value function with Q-values as a state by action matrix.}

\item{epsilon}{an \code{epsilon > 0} applies an epsilon-greedy policy.}

\item{prob}{probability vector for random actions for random_MDP_policy().
a logical indicating if action probabilities should be returned for
\code{greedy_MDP_action()}.}

\item{actions}{a vector with the action (either the action label or the
numeric id) for each state.}
}
\value{
\code{q_values_MDP()} returns a state by action matrix specifying the Q-function,
i.e., the action value for executing each action in each state. The Q-values
are calculated from the value function (U) and the transition model.

\code{approx_MDP_policy_evaluation()} returns a vector with approximate
state values (U).

\code{greedy_MDP_action()} returns the action with the highest q-value
for state \code{s}. If \code{prob = TRUE}, then a vector with
the probability for each action is returned.

\code{greedy_MDP_policy()} returns the greedy policy given \code{Q}.

\code{random_MDP_policy()} returns a data.frame with the columns state and action to define a policy.

\code{manual_MDP_policy()} returns a data.frame with the columns state and action to define a policy.
}
\description{
Implementation several functions useful to deal with MDP policies.
}
\examples{
data(Maze)
Maze

# create several policies:
# 1. optimal policy using value iteration
maze_solved <- solve_MDP(Maze, method = "value_iteration")
maze_solved
pi_opt <- policy(maze_solved)
pi_opt
plot_Maze_solution(add_policy(Maze, pi_opt), main = "Optimal Policy")

# 2. a manual policy (go up and in some squares to the right)
acts <- rep("up", times = length(Maze$states))
acts[c(3, 6, 9)] <- "right"
pi_manual <- manual_MDP_policy(Maze, acts)
pi_manual
plot_Maze_solution(add_policy(Maze, pi_manual), main = "Manual Policy")

# 3. a random policy
pi_random <- random_MDP_policy(Maze)
pi_random
plot_Maze_solution(add_policy(Maze, pi_random), main = "Random Policy")

# 4. an improved policy based on one policy evaluation and
#   policy improvement step
u <- approx_MDP_policy_evaluation(pi_random, Maze)
q <- q_values_MDP(Maze, U = u)
pi_greedy <- greedy_MDP_policy(q)
pi_greedy
plot_Maze_solution(add_policy(Maze, pi_greedy), , main = "Greedy Policy")

#' compare the approx. value functions for the policies
rbind(
  random = approx_MDP_policy_evaluation(pi_random, Maze),
  manual = approx_MDP_policy_evaluation(pi_manual, Maze),
  greedy = approx_MDP_policy_evaluation(pi_greedy, Maze),
  optimal = approx_MDP_policy_evaluation(pi_opt, Maze)
)


# For many functions, we first add the policy to the problem description
#   to create a "solved" MDP
maze_random <- add_policy(Maze, pi_random)
maze_random

# plotting
plot_value_function(maze_random)
plot_Maze_solution(maze_random)

# compare to a benchmark
regret(maze_random, maze_solved, start = "s_1")

# calculate greedy actions for state 1
q <- q_values_MDP(Maze_random)
q
greedy_MDP_action(1, q, epsilon = 0, prob = FALSE)
greedy_MDP_action(1, q, epsilon = 0, prob = TRUE)
greedy_MDP_action(1, q, epsilon = .1, prob = TRUE)
}
\seealso{
Other MDP: 
\code{\link{MDP}()},
\code{\link{accessors}},
\code{\link{add_policy}()},
\code{\link{regret}()},
\code{\link{simulate_MDP}()},
\code{\link{solve_MDP}()},
\code{\link{transition_graph}()},
\code{\link{value_function}()}
}
\concept{MDP}
