% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/policy_graph.R
\name{policy_graph}
\alias{policy_graph}
\title{Extract the Policy Graph (as an igraph Object)}
\usage{
policy_graph(x, belief = TRUE, col = NULL, ...)
}
\arguments{
\item{x}{A \link{POMDP} object.}

\item{belief}{logical; calculate belief proportions and show them as a pie chart in each node of
the graph?}

\item{col}{colors used for the states in the belief proportions.}

\item{...}{if \code{belief = TRUE}, then additional parameters are passed on for belief point sampling with \code{\link[=sample_belief_space]{sample_belief_space()}} or
\code{\link[=simulate_POMDP]{simulate_POMDP()}}.}
}
\value{
An object of class igraph containing a directed graph.
}
\description{
Convert the policy graph in a POMDP solution object into an igraph object.
}
\examples{
data("Tiger")
sol <- solve_POMDP(model = Tiger)
sol

pg <- policy_graph(sol)

plot(pg)

}
\seealso{
Other policy: 
\code{\link{optimal_action}()},
\code{\link{plot_policy_graph}()},
\code{\link{plot_value_function}()},
\code{\link{policy}()},
\code{\link{reward}()},
\code{\link{solve_POMDP}()},
\code{\link{solve_SARSOP}()}
}
\author{
Hossein Kamalzadeh, Michael Hahsler
}
\concept{policy}
\keyword{graphs}
