% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AAA_pomdp-package.R
\docType{package}
\name{pomdp-package}
\alias{pomdp-package}
\title{pomdp: Solver for Partially Observable Markov Decision Processes (POMDP)}
\description{
Provides the infrastructure to define and analyze the solutions of Partially Observable Markov Decision Process (POMDP) models. Interfaces for various exact and approximate solution algorithms are available including value iteration, Point-Based Value Iteration (PBVI) and Successive Approximations of the Reachable Space under Optimal Policies (SARSOP).
}
\section{Key functions}{

\itemize{
\item Problem specification: \link{POMDP}, \link{MDP}
\item Solvers: \code{\link[=solve_POMDP]{solve_POMDP()}}, \code{\link[=solve_MDP]{solve_MDP()}}, \code{\link[=solve_SARSOP]{solve_SARSOP()}}
}
}

\author{
Michael Hahsler
}
