% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AAA_pomdp-package.R
\docType{package}
\name{pomdp-package}
\alias{pomdp-package}
\title{pomdp: Solver for Partially Observable Markov Decision Processes (POMDP)}
\description{
Provides the infrastructure to define and analyze the solutions of Partially Observable Markov Decision Processes (POMDP) models solved using pomdp-solve or SARSOP using a variety of exact and approximate value iteration algorithms.
}
\section{Key functions}{

\itemize{
\item \code{\link[=solve_POMDP]{solve_POMDP()}}, \code{\link[=solve_MDP]{solve_MDP()}}
}
}

\author{
Michael Hahsler
}
