\name{solver_output}
\alias{solver_output}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Display the Output of the POMDP Solver
}
\description{
Displays the output generated by the solver 'pomdp-solve'. This includes used parameters,
and iterations (i.e., epochs). This produces the same output as running \code{solve_POMDP}
with the argument \code{verbose = TRUE}.
}
\usage{
solver_output(x)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{x}{
object of class POMDP returned by \code{\link{solve_POMDP}}.
}
}

\value{
returns invisibly a character string vector with the output of 'pomdp-solve'.
}

\seealso{
\code{\link{solve_POMDP}}
}
\examples{
data("TigerProblem")
sol <- solve_POMDP(model = TigerProblem)

## solver output
solver_output(sol)
}
\keyword{optimize}