% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tiger.R
\docType{data}
\name{Tiger}
\alias{Tiger}
\alias{Three_doors}
\title{Tiger Problem POMDP Specification}
\format{
An object of class \link{POMDP}.
}
\description{
The model for the Tiger Problem was introduced by Cassandra et al. (1994).
}
\details{
The original Tiger problem was published in Cassandra et al (1994) as
follows:

An agent is facing two closed doors and a tiger is put with equal
probability behind one of the two doors represented by the states
\code{tiger-left} and \code{tiger-right}, while treasure is put behind the other door.
The possible actions are \code{listen} for tiger noises or opening a door (actions
\code{open-left} and \code{open-right}). Listening is neither free (the action has a
reward of -1)  nor is it entirely accurate. There is a 15\\% observation
probability that the agent hears the tiger behind the left door while it is
actually behind the right door and vice versa. If the agent opens the door with
the tiger, it will get hurt (a negative reward of -100), but if it opens the
door with the treasure, it will receive a positive reward of 10. After a door
is opened, the problem is reset (i.e., the tiger is randomly assigned to a
door with a 50/50 chance), and the agent gets another try.

The three doors problem is an extension of the Tiger problem where the tiger
is behind one of three doors represented by three states (\code{tiger-left},
\code{tiger-center}, and \code{tiger-right}) and treasure is behind the other two
doors. There are also three open actions and three different observations for
listening.
}
\examples{
data("Tiger")
Tiger

data("Three_doors")
Three_doors
}
\references{
Anthony R. Cassandra, Leslie P Kaelbling, and Michael L.
Littman (1994). Acting Optimally in Partially Observable Stochastic Domains.
In Proceedings of the Twelfth National Conference on Artificial
Intelligence, pp. 1023-1028.
}
\seealso{
Other POMDP_examples:
\code{\link[=POMDP]{POMDP()}},
\code{\link{POMDP_example_files}},
\code{\link{RussianTiger}}
}
\concept{POMDP_examples}
\keyword{datasets}
