% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transition_matrix.R
\name{transition_matrix}
\alias{transition_matrix}
\alias{transition_function}
\alias{observation_matrix}
\alias{observation_function}
\alias{reward_matrix}
\alias{reward_function}
\title{Extract the Transition, Observation or Reward Information from a POMDP}
\usage{
transition_matrix(x, episode = 1, action = NULL)

transition_function(x, episode = 1)

observation_matrix(x, episode = 1, action = NULL)

observation_function(x, episode = 1)

reward_matrix(x, episode = 1, action = NULL)

reward_function(x, episode = 1)
}
\arguments{
\item{x}{A \link{POMDP} object.}

\item{episode}{Episode used for time-dependent POMDPs (\link{POMDP}).}

\item{action}{only return the matrix for a given action.}
}
\value{
A list or a list of lists of matrices.
}
\description{
Converts the description of transition probabilities and observation
probabilities in a POMDP into a list of matrices or a function.
}
\details{
See Details section in \link{POMDP} for details.
}
\examples{
data("Tiger")

# List of |A| transition matrices. One per action in the from states x states
Tiger$transition_prob
transition_matrix(Tiger)

f <- transition_function(Tiger)
args(f)
## listening does not change the tiger's position.
f("listen", "tiger-left", "tiger-left")

# List of |A| observation matrices. One per action in the from states x observations
Tiger$observation_prob
observation_matrix(Tiger)

# List of list of reward matrices. 1st level is action and second level is the
#  start state in the form end state x observation
Tiger$reward
reward_matrix(Tiger)

# Visualize transition matrix for action 'open-left'
library("igraph")
g <- graph_from_adjacency_matrix(transition_matrix(Tiger)$"open-left", weighted = TRUE)
edge_attr(g, "label") <- edge_attr(g, "weight")

igraph.options("edge.curved" = TRUE)
plot(g, layout = layout_on_grid, main = "Transitions for action 'open=left'")

## Use a function for the Tiger transition model
trans <- function(action, end.state, start.state) {
  ## listen has an identity matrix
  if(action == 'listen')
    if(end.state == start.state) return(1)
    else return(0)

  # other actions have a uniform distribution
  return(1/2)
}

Tiger$transition_prob <- trans
transition_matrix(Tiger)
}
\seealso{
Other POMDP: 
\code{\link{POMDP}()},
\code{\link{plot_belief_space}()},
\code{\link{sample_belief_space}()},
\code{\link{simulate_POMDP}()},
\code{\link{solve_POMDP}()},
\code{\link{solve_SARSOP}()},
\code{\link{update_belief}()},
\code{\link{write_POMDP}()}
}
\author{
Michael Hahsler
}
\concept{POMDP}
